refactor tests to separate bdf testing module
